Move db into a separate crate
