+++ Cargo.toml
@@ -12,22 +12,16 @@ axum = { version = "0.6.11", features = ["headers", "json"] }
 axum-server = "0.4.6"
 chrono = "0.4.24"
 clap = { version = "4.1.8", features = ["derive"] }
+db = { path = "crates/db" }
 env_logger = "0.10.0"
 extractor = { path = "crates/extractor" }
 futures = "0.3.27"
 http = "0.2.9"
-itertools = "0.10.5"
 log = "0.4.17"
 model = { path = "crates/model" }
-mongodb = "2.4.0"
 rayon = "1.7.0"
 reqwest = { version = "0.11.14", features = ["blocking"] }
 serde = { version = "1.0.156", features = ["derive"] }
 serde_json = "1.0.94"
 tokio = { version = "1.26.0", features = ["rt-multi-thread", "macros"] }
 tower-http = { version = "0.4.0", features = ["cors"] }
-
-[dev-dependencies]
-include_dir = "0.7.3"
-pretty_assertions = "1.3.0"
-tempdir = "0.3.7"
