+++ crates/db/Cargo.toml
@@ -1,0 +1,20 @@
+[package]
+name = "db"
+version = "0.0.0"
+edition = "2021"
+publish = false
+
+[dependencies]
+anyhow = "1.0.70"
+futures = "0.3.27"
+itertools = "0.10.5"
+log = "0.4.17"
+model = { path = "../../crates/model" }
+mongodb = "2.4.0"
+serde_json = "1.0.94"
+tokio = { version = "1.26.0", features = ["rt-multi-thread", "macros"] }
+
+[dev-dependencies]
+include_dir = "0.7.3"
+pretty_assertions = "1.3.0"
+tempdir = "0.3.7"
