+++ crates/db/src/lib.rs
@@ -1,0 +1,28 @@
+use {
+  crate::vec_ext::VecExt,
+  itertools::Itertools,
+  model::Course,
+  mongodb::{
+    bson::{doc, Document},
+    options::ClientOptions,
+    options::UpdateModifications,
+    results::{InsertOneResult, UpdateResult},
+    Client, Database,
+  },
+  std::{fs, hash::Hash, path::PathBuf},
+};
+
+#[cfg(test)]
+use {
+  futures::stream::TryStreamExt,
+  include_dir::{include_dir, Dir},
+  std::sync::atomic::{AtomicUsize, Ordering},
+  tempdir::TempDir,
+};
+
+type Result<T = (), E = anyhow::Error> = std::result::Result<T, E>;
+
+mod db;
+mod vec_ext;
+
+pub use crate::db::Db;
