+++ src/vec_ext.rs
@@ -1,18 +1,12 @@
 use super::*;
 
 pub(crate) trait VecExt<T> {
-  fn combine(self, other: Vec<T>) -> Vec<T>;
-
   fn into_option(self) -> Option<Self>
   where
     Self: Sized;
 }
 
-impl<T: Eq + Hash + Clone> VecExt<T> for Vec<T> {
-  fn combine(self, other: Vec<T>) -> Vec<T> {
-    [self, other].concat().iter().unique().cloned().collect()
-  }
-
+impl<T> VecExt<T> for Vec<T> {
   fn into_option(self) -> Option<Self> {
     match self.is_empty() {
       true => None,
